LayoutTests/storage/filesystemaccess/file-handle-getfile-expected.txt

This test checks getFile() of FileSystemFileHandle.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS readText is fileContent
PASS successfullyParsed is true

TEST COMPLETE
